python/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0

# Deep Learning (optional)
torch>=2.0.0
transformers>=4.20.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
bokeh>=2.4.0

# Time Series
statsmodels>=0.13.0

# Dimensionality Reduction
umap-learn>=0.5.0

# Natural Language Processing
nltk>=3.7.0
spacy>=3.4.0

# Utilities
tqdm>=4.64.0
joblib>=1.1.0

# Clustering
hdbscan>=0.8.28

# ===== Phase 7-9: Production & Advanced ML =====
# Memory & Performance
psutil>=5.9.0

# Cloud & Database
boto3>=1.26.0  # AWS S3
azure-storage-blob>=12.14.0  # Azure Blob
google-cloud-storage>=2.7.0  # Google Cloud Storage
pymongo>=4.3.0  # MongoDB
psycopg2-binary>=2.9.5  # PostgreSQL
PyMySQL>=1.0.2  # MySQL

# Time Series
prophet>=1.1.0
pmdarima>=2.0.0

# NLP
vaderSentiment>=3.3.2
textblob>=0.17.1

# Model Interpretability
shap>=0.41.0

# ===== Phase 10: MLOps & Deployment =====
# API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
slowapi>=0.1.8  # Rate limiting

# MLOps
mlflow>=2.5.0
evidently>=0.4.0  # Model monitoring
alibi-detect>=0.11.0  # Drift detection

# Docker
docker>=6.1.0
python-on-whales>=0.64.0

# Advanced NLP
gensim>=4.3.0  # Topic modeling
bertopic>=0.15.0  # BERT-based topics
pyLDAvis>=3.4.0  # Topic visualization
sentence-transformers>=2.2.0  # Document similarity
faiss-cpu>=1.7.0  # Similarity search

# Dashboards
streamlit>=1.25.0
dash>=2.11.0

# Monitoring & Metrics
prometheus-client>=0.17.0

# ===== Notebook to Pipeline =====
# Jupyter notebook parsing
nbformat>=5.7.0
nbconvert>=7.0.0